+++ classes/person_lister.rb
@@ -1,4 +1,4 @@
-require_relative './person'
+require_relative 'person'
 
 class PersonLister
   def initialize(people)
