+++ classes/utility/menu.rb
@@ -1,0 +1,12 @@
+class Menu
+  def print
+    puts 'What would you like to do?'
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals'
+    puts '7 - Quit'
+  end
+end
